HomeWork#5

+++ MetricsAgent/DAL/Support/DateTimeOffsetHandler.cs
@@ -12,6 +12,6 @@ namespace MetricsAgent.DAL
         public override void SetValue(IDbDataParameter parameter, DateTimeOffset value)
             => parameter.Value = value;
 
-        public override DateTimeOffset Parse(object value) => DateTimeOffset.FromUnixTimeSeconds((int)value);
+        public override DateTimeOffset Parse(object value) => DateTimeOffset.FromUnixTimeSeconds((long)value);
     }
 }
